renamed uploader-middleware.js to s3-middleware.js
